+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion "25.0.3"
     defaultConfig {
-        applicationId "vola.systers.com.volunteers_android"
+        applicationId "vola.systers.com.android"
         minSdkVersion 15
         targetSdkVersion 25
         versionCode 1
