+++ app/src/main/java/vola/systers/com/android/fragments/StarredEventsFragment.java
@@ -1,4 +1,4 @@
-package vola.systers.com.volunteers_android.fragments;
+package vola.systers.com.android.fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,21 +6,20 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import vola.systers.com.volunteers_android.R;
+import vola.systers.com.android.R;
 
-public class NavigateFragment extends Fragment {
+public class StarredEventsFragment extends Fragment {
 
-    public NavigateFragment(){}
+    public StarredEventsFragment(){}
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        // TODO : NAVIGATE FRAGMENT
+        //TODO : STARRED EVENTS FRAGMENT
 
         View rootView = inflater.inflate(R.layout.profile_fragment, container, false);
         return rootView;
     }
-
 
 }
