+++ app/src/main/java/vola/systers/com/android/utils/HttpHandler.java
@@ -1,4 +1,4 @@
-package vola.systers.com.volunteers_android.handler;
+package vola.systers.com.android.utils;
 
 import android.util.Log;
 
@@ -19,7 +19,7 @@ public class HttpHandler {
     public HttpHandler() {
     }
 
-    public String makeServiceCall(String reqUrl) {
+    public static String makeServiceCall(String reqUrl) {
         String response = null;
         try {
             URL url = new URL(reqUrl);
@@ -40,7 +40,7 @@ public class HttpHandler {
         return response;
     }
 
-    private String convertStreamToString(InputStream is) {
+    private static String convertStreamToString(InputStream is) {
         BufferedReader reader = new BufferedReader(new InputStreamReader(is));
         StringBuilder sb = new StringBuilder();
